refactored weather DTOs and migrated HomeScreen to Android ViewModel with Koin integration

+++ app/src/main/java/com/example/weather/HomeScreen/UI_Presentation/component/MainHeaderTemp.kt
@@ -44,7 +44,7 @@ private fun MainHeaderTempPreview() {
         region = "i dont know",
         country = "Germany",
         localtimeEpoch = 323013200,
-        temperatureC = 32,
+        temperatureC = 32.toFloat(),
         conditionText = "Light rain",
         conditionIconUrl = "er",
         windKph = 23F,
